+++ test/nextgen/operators/matmul/pack_lhs/matmul_pack_lhs_wrapper.hpp
@@ -34,6 +34,10 @@ namespace kai::test {
 [[nodiscard]] bool is_shape_suitable_lhs_qai8dxp1vlx8_qsi4cxp4vlx8_1vlx4vl_sme2_mopa(
     size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
 
+/// Checks if the portion produces non-empty LHS packing tiles for the qai8dxp1vlx8/qsi4cxp4vlx8 matmul operator.
+[[nodiscard]] bool is_shape_suitable_lhs_qai8dxp1vlx4_qsi4cxp4vlx4_1vlx4vl_sme_mopa(
+    size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
+
 /// Checks if the portion produces non-empty LHS packing tiles for the qai8dxp1x4/qsi4cxp4vlx4 matmul operator.
 [[nodiscard]] bool is_shape_suitable_lhs_qai8dxp1x4_qsi4cxp4vlx4_1x4vl_sme2_sdot(
     size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);